requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.2.0
alibabacloud-credentials==0.3.0
alibabacloud-darabonba-array==0.1.0
alibabacloud-darabonba-encode-util==0.0.1
alibabacloud-darabonba-map==0.0.1
alibabacloud-darabonba-signature-util==0.0.3
alibabacloud-darabonba-string==0.0.4
alibabacloud-ecs20140526==3.0.1
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.1
alibabacloud-openapi-util==0.2.0
alibabacloud-ros20190910==3.2.16
alibabacloud-tea==0.3.0
alibabacloud-tea-openapi==0.3.6
alibabacloud-tea-util==0.3.8
alibabacloud-tea-xml==0.0.2
alibabacloud-vpc20160428==2.0.12
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backports.shutil-get-terminal-size==1.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
crcmod==1.7
cryptography==38.0.3
dataclasses-jsonschema==2.16.0
debtcollector==2.5.0
decorator==5.1.1
dulwich==0.20.50
exceptiongroup==1.0.4
frozenlist==1.3.1
idna==3.4
iniconfig==1.1.1
iso8601==1.1.0
jmespath==0.10.0
jsonschema==4.17.0
multidict==6.0.2
netaddr==0.8.0
netifaces==0.11.0
oslo.i18n==5.1.0
oslo.utils==6.0.1
oss2==2.16.0
packaging>=21.3
pbr==5.11.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pycryptodome==3.16.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
reprint==0.6.0
requests==2.28.1
retrying==1.3.4
six==1.16.0
tabulate==0.9.0
tomli==2.0.1
urllib3==1.26.12
wrapt==1.14.1
yarl==1.8.1
yattag==1.14.0
